# src/docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorDetails:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        description: ESTE NO IRIA MAS
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  errors.ErrorDetailsWithParams:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
      validation_errors:
        items:
          $ref: '#/definitions/errors.ErrorParam'
        type: array
    type: object
  errors.ErrorParam:
    properties:
      field:
        type: string
      reason:
        type: string
    type: object
  model.CommunityResponse:
    properties:
      description:
        type: string
      id:
        type: string
      joined:
        type: boolean
      name:
        type: string
      owner_id:
        type: string
    type: object
  model.NewCommunityPostRequest:
    properties:
      content:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - title
    type: object
  model.NewCommunityRequest:
    properties:
      description:
        type: string
      name:
        type: string
      picture:
        items:
          type: integer
        type: array
    required:
    - description
    - name
    type: object
...
  models.UserStageResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      id_register:
        type: string
      is_author:
        type: boolean
      last_name:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a  server for Swagger with Gin.
  title: BetterReads API
  version: "1.0"
paths:
...
  /communities:
    get:
      consumes:
      - application/json
      description: Get all communities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CommunityResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all communities
      tags:
      - communities
    post:
      consumes:
      - multipart/form-data
      description: creates a community, the community data should follow the model.NewCommunityRequest
        in JSON
      parameters:
      - description: Community Data
        in: formData
        name: data
        required: true
        type: string
      - description: Community Picture
        in: formData
        name: file
        required: true
        type: file
      - description: Don't need to send this in json, this param is only here to reference
          NewCommunityRequest, DONT SEND PICTURE in JSON
        in: body
        name: community
        required: true
        schema:
          $ref: '#/definitions/model.NewCommunityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.CommunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetailsWithParams'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: creates a community
      tags:
      - communities
  /communities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a community if the user is the creator
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      security:
      - ApiKeyAuth: []
      summary: Delete a community
      tags:
      - communities
    get:
      consumes:
      - application/json
      description: Get a community by id
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommunityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get a community by id
      tags:
      - communities
  /communities/{id}/join:
    post:
      consumes:
      - application/json
      description: Join a community
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Join a community
      tags:
      - communities
  /communities/{id}/leave:
    delete:
      consumes:
      - application/json
      description: Leave a community
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Leave a community
      tags:
      - communities
  /communities/{id}/picture:
    get:
      description: Get community picture
      parameters:
      - description: Community id
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorDetails'
      summary: Get Community picture
      tags:
      - communities
  /communities/{id}/posts:
    get:
      consumes:
      - application/json
      description: gets the posts of a community
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Post Content
        in: body
        name: content
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: gets the posts of a community
      tags:
      - communities
    post:
      consumes:
      - application/json
      description: Create a post in a community
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      - description: Post Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.NewCommunityPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Create a post in a community
      tags:
      - communities
  /communities/{id}/users:
    get:
      consumes:
      - application/json
      description: Get all users in a community
      parameters:
      - description: Community ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserStageResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all users in a community
      tags:
      - communities
  /communities/search:
    get:
      consumes:
      - application/json
      description: Search communities by name. It returns all communities that have
        the search string in their name
      parameters:
      - description: Search string
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CommunityResponse'
            type: array
      summary: Search communities by name
      tags:
      - communities
...
swagger: "2.0"